requirements.txt
cvxpy==1.5.1
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.1
scipy==1.10.1